+++ src/lib/addVideo.ts
@@ -16,6 +16,7 @@ export async function addVideo(profile: string): Promise<void> {
 	const title      = await getTitle(cli);
 	const duration   = await getDuration(cli);
 	const resolution = await getResolution(cli);
+	const epoch      = await getEpoch(cli);
 
 	const videoFile       = await getVideoFile(cli);
 	const imageFile       = await getImageFile(cli);
@@ -29,7 +30,7 @@ export async function addVideo(profile: string): Promise<void> {
 	fs.ensureDir(outputDir, { create: true });
 
 	const filename = path.join(outputDir, `${formatTitle({ id, title, channel }).toFilename()}`);
-	const json     = formatJSON({ id, title, channel, ext, resolution, duration });
+	const json     = formatJSON({ id, title, channel, ext, resolution, duration, epoch });
 
 	await fs.promises.writeFile(`${filename}.info.json`, JSON.stringify(json));
 
@@ -87,6 +88,21 @@ async function getResolution(cli: Cli): Promise<[number, number]> {
 		}
 
 		return match.slice(1).map((d) => parseInt(d)).toTuple(2);
+	});
+}
+
+async function getEpoch(cli: Cli): Promise<number> {
+	const format = 'YYYY-MM-dd';
+	const regex  = /^(\d{4})-(\d{2})-(\d{2})$/;
+
+	return cli.getAnswer('Date', (answer) => {
+		const match = answer.match(regex);
+
+		if (!match) {
+			return new Error(`Date should be in the format ${format}`);
+		}
+
+		return Math.round(new Date(match[1]!).getTime() / 1000);
 	});
 }