+++ src/lib/types/index.ts
@@ -1,13 +1,6 @@
-export interface VideoInfo {
-	id: string;
-	title: string;
-	channel: string;
-	ext: string;
-}
+import type z from 'zod';
 
-export interface VideoJSON extends VideoInfo {
-	width: number;
-	height: number;
-	resolution: string;
-	duration_string: string;
-}
+import type { videoInfoSchema, videoJSONSchema } from './schema';
+
+export type VideoInfo = z.infer<typeof videoInfoSchema>;
+export type VideoJSON = z.infer<typeof videoJSONSchema>;