+++ src/lib/utils/formatTitle.ts
@@ -1,5 +1,5 @@
 import type { VideoInfo } from '../types';
 
-export function formatTitle({ id, title, channel }: Omit<VideoInfo, 'ext'>): string {
+export function formatTitle({ id, title, channel }: Pick<VideoInfo, 'id' | 'title' | 'channel'>): string {
 	return `${title} [${channel}].${id}`;
 }