+++ src/lib/validate.ts
@@ -2,8 +2,9 @@ import fs from 'fs';
 import path from 'path';
 
 import { log, warn } from '@anmiles/logger';
+import { validate as zodValidate } from '@anmiles/zod-tools';
 
-import type { VideoInfo } from './types';
+import { videoInfoSchema } from './types/schema';
 import { formatTitle } from './utils/formatTitle';
 import { getOutputDir } from './utils/paths';
 
@@ -26,10 +27,10 @@ export function validate(profile: string): void {
 			file.exts.push(ext);
 
 			if (ext === '.info.json') {
-				// TODO: zod
 				// TODO: fs - replace readJSON "as" with zod validation
-				const info   = fs.readJSON<VideoInfo>(filepath);
-				file.newName = formatTitle(info).toFilename();
+				const infoObject = fs.readJSON(filepath);
+				const info       = zodValidate(infoObject, videoInfoSchema);
+				file.newName     = formatTitle(info).toFilename();
 			}
 		},
 	}, { depth: 1 });