+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [11.2.0](../../tags/v11.2.0) - 2026-02-03
+### Add
+- Write down date when adding video manually
 ### Changed
 - Add `--js-runtimes` argument to the executable command