+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [11.3.0](../../tags/v11.3.0) - 2026-02-21
+### Add
+- Ability to manually input data as JSON
+
 ## [11.2.0](../../tags/v11.2.0) - 2026-02-03
 ### Add
 - Write down date when adding video manually